Add files via upload
Version 1.0.  If you have Node.js installed, download all this to a directory, enter it, type "npm install" to get the dependencies, then "node server" or "node server.js" (the extension is optional).  Connect on port 3000, as in start up a browser and point it to http://localhost:3000/ and you should see the game.  Once a second player connects (or the first player twice, if you're testing), it will let you start a game.  Have fun!